Improve head test with better array validation
This is probably not the best solution, but the better solution
(to save the array name rather than its pointer) would require
refactoring the pwb_handle implementation and would take a
performance hit to search the variables hash table for every
access to a handle.

The solution checks the ate_header::array as an int to see
if the address value is a repeating bit pattern, which likely
indicates that the variable has been freed.  The follow-up
test, if the repeating pattern is not detected, is to confirm
the the address of the ARRAY_ELEMENT is the same as the
ARRAY::head::next value.

The error I'm protecting against is if the resident array
goes out-of-scope and subsequently freed  before the handle
is used.  Even if it's not a perfect test, it's good enough
to prompt the developer to fix it before a fluke false negative
causes a seg-fault.