Remove row-count/array size comparisons when testing head integrity
With something, something, get_handle_or_die, I'm now calling function
to test head integrity.  Unfortunately, the test is necessary to protect
against an seg-fault for an out-of-scope array, but the test was also
checking for an out-of-date index.  append_data and index_rows (among
others) necessarily result in an out-of-date index, so I had to remove
the test.